Configクラス定義を分離
